bha_selection/requirements.txt
# BHA Selection Tool - Dependencies
requests>=2.31
python-dotenv>=1.0
pandas>=2.1
pyarrow>=15.0
matplotlib>=3.8
fastapi>=0.100
uvicorn>=0.27
pydantic>=2.5
